requirements.txt
Django==4.1.2
django-allauth==0.50.0
Pillow==9.5.0
django-filter==21.1
djangorestframework==3.14.0